// apps/api/src/modules/auth/dto/auth.dto.d.ts
export declare class RegisterDto {
    name: string;
    email: string;
    password: string;
}
export declare class LoginDto {
    email: string;
    password: string;
}
export declare class RefreshTokenDto {
    refreshToken: string;
}
export declare class LogoutDto {
    refreshToken?: string;
}
//# sourceMappingURL=auth.dto.d.ts.map